Refactor Tracker module structure and update imports
Moved Tracker-related classes into a new 'src' subpackage, updated package declarations and imports accordingly, and removed the old TrackerModel implementation. Introduced FileInfoAdapter for Gson serialization, updated method and field names for consistency, and improved data structure usage for file and peer management. This refactor clarifies module boundaries and prepares the codebase for further enhancements.
