+++ src/Tracker/src/model/ClientState.java
@@ -1,4 +1,4 @@
-package model;
+package src.model;
 
 import java.nio.ByteBuffer;
