+++ src/Tracker/src/model/RequestHandler.java
@@ -1,4 +1,4 @@
-package model;
+package src.model;
 
 public interface RequestHandler {
     String handle(String[] parts);
