+++ src/Tracker/src/utils/Log.java
@@ -1,4 +1,4 @@
-package utils;
+package src.utils;
 
 public class Log {
     public static void logInfo(String message) {
