+++ src/backend_layer/src/main/java/domain/repositories/FileRepository.java
@@ -11,21 +11,34 @@ import java.util.Set;
 
 public interface FileRepository {
     void shareFileAsync(File file, String fileName, String progressId, int isReplace, FileInfo oldFileInfo);
+
     void shareFileList();
+
     void downloadFile(FileInfo fileInfo, File saveFile, List<PeerInfo> peers, String progressId);
+
     void resumeDownload(String progressId);
+
     List<PeerInfo> getPeersWithFile(String fileHash);
+
     boolean shareFileToPeers(File file, FileInfo oldFileInfo, int isReplace, String progressId, List<String> peerList);
-    List<String> getSelectivePeers(String fileHash);
-    List<String> getKnownPeers();
+
     void loadSharedFiles();
+
     int refreshSharedFileNames();
+
+    List<String> getKnownPeers();
+
     Set<FileInfo> getSharedFileNames();
-    void setSharedFileNames(Set<FileInfo> sharedFileNames);
+
     Map<String, FileInfo> getPublicSharedFiles();
-    Map<String, FileInfo> getPrivateSharedFiles();
+
+    Map<FileInfo, Set<PeerInfo>> getPrivateSharedFiles();
+
     int stopSharingFile(String fileName);
+
     Map<String, ProgressInfo> getProgress();
+
     void setProgress(ProgressInfo progressInfo);
+
     void cleanupProgress(List<String> progressIds);
 }
