+++ src/backend_layer/src/main/java/utils/Infor.java
@@ -7,8 +7,8 @@ package main.java.utils;
 public interface Infor {
     int SERVER_PORT = EnvConf.getEnvInt("SERVER_PORT", 5000);
     int TRACKER_PORT = EnvConf.getEnvInt("TRACKER_PORT", 5001);
-    String SERVER_IP = GetLocalIP.getCurrentIp();
-    String TRACKER_IP = GetLocalIP.getCurrentIp();
+    String SERVER_IP = NetworkUtils.getCurrentIp();
+    String TRACKER_IP = NetworkUtils.getCurrentIp();
     int CHUNK_SIZE = EnvConf.getEnvInt("CHUNK_SIZE", 1024 * 1024 * 2);
     int SOCKET_TIMEOUT_MS = EnvConf.getEnvInt("SOCKET_TIMEOUT_MS", 5000);
     int DOWNLOAD_TIMEOUT_MS = 10000;
